refactor: assemble hooks in single file
