+++ lib/createInlinePluginCreator.js
@@ -1,11 +1,9 @@
 const { writeFileSync } = require("fs");
-// const { resolve } = require("path");
 const { check } = require("./blork");
 const wait = require("./wait");
 const getCommitsFiltered = require("./getCommitsFiltered");
 const getManifest = require("./getManifest");
 const hasChangedDeep = require("./hasChangedDeep");
-const debug = require("./debug");
 
 /**
  * Create an inline plugin creator for a multirelease.
