+++ lib/getLogger.js
@@ -1,5 +1,5 @@
 const { Signale } = require("signale");
-const once = require("lodash.once");
+const { once } = require("lodash");
 
 /**
  * Return a new Signale instance.
