+++ package.json
@@ -52,12 +52,9 @@
 		"blork": "^9.2.2",
 		"cosmiconfig": "^6.0.0",
 		"execa": "^4.0.0",
+		"lodash": "^4.17.15",
 		"get-stream": "^5.1.0",
 		"git-log-parser": "^1.2.0",
-		"lodash.get": "^4.4.2",
-		"lodash.once": "^4.1.1",
-		"lodash.pairs": "^3.0.1",
-		"lodash.set": "^4.3.2",
 		"meow": "^6.0.1",
 		"require-in-the-middle": "^5.0.3",
 		"semantic-release": "^17.0.4",
